+++ models/sectionTypes/TextSimple.php
@@ -94,8 +94,8 @@ class TextSimple extends ISectionType
             $str .= $amParas[$paraNo];
             $str .= '</div>';
 
-            $str .= '<div class="modifiedActions"><a href="#" class="revert">';
-            $str .= \Yii::t('amend', 'singlepara_revert');
+            $str .= '<div class="modifieActions"><a href="#" class="revert">';
+            $str .= \Yii::t('amend', 'revert_changes');
             $str .= '</a></div>';
 
             $str .= '</div>';