+++ tests/acceptance/admin/MotionListMotionAsTemplateCept.php
@@ -34,18 +34,3 @@ $name2 = $I->executeJS('return $(".supporterData .supporterRow").eq(1).find("inp
 if ($name1 != 'Felix Deist' || $name2 != 'Tarek Al-Wazir') {
     $I->fail('supporter data not present');
 }
-
-
-$I->gotoMotionList();
-$I->click('.adminMotionTable .amendment13 .actionCol .dropdown-toggle');
-$I->click('.adminMotionTable .amendment13 .actionCol .asTemplate');
-$I->see('Änderungsantrag', 'h1');
-$I->seeInField('#initiatorPrimaryName', 'Daniel Gollasch');
-$I->dontSeeElement('#resolutionDate');
-$I->seeCheckboxIsChecked('#personTypeNatural');
-$I->seeCheckboxIsChecked('input[name=otherInitiator]');
-$name1 = $I->executeJS('return $(".supporterData .supporterRow").eq(0).find("input.name").val()');
-$name2 = $I->executeJS('return $(".supporterData .supporterRow").eq(1).find("input.name").val()');
-if ($name1 != 'Antje Kapek, LV Berlin' || $name2 != 'Axel Vogel, LV Brandenburg') {
-    $I->fail('supporter data not present');
-}